requirements.txt
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cssselect2==0.7.0
Django==4.1.3
django-appconf==1.0.5
django-image-cropping==1.7
django-paypal==2.0
django-phonenumber-field==7.0.0
django-twilio==0.14.3
easy-thumbnails==2.8.3
fonttools==4.38.0
html5lib==1.1
idna==3.4
numpy==1.23.5
pandas==1.5.2
phonenumbers==8.13.0
Pillow==9.3.0
plotly==5.6.0
psycopg2==2.9.5
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
pyphen==0.13.2
python-dateutil==2.8.2
pytz==2022.6
razorpay==1.3.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tenacity==8.1.0
tinycss2==1.2.1
twilio==7.15.2
tzdata==2022.6
urllib3==1.26.12
weasyprint==57.1
webencodings==0.5.1
zopfli==0.2.2
